Make usage of the generic Write method...

+++ Projects/Framework/Packets/Server/Character/CharacterLoginFailed.cs
@@ -29,7 +29,7 @@ namespace Framework.Packets.Server.Character
 
         public override void Write()
         {
-            Packet.Write((byte)Code);
+            Packet.Write(Code);
         }
     }
 }
